+++ CHANGELOG.md
@@ -1,9 +1,12 @@
 # Changelog
 
-## 0.0.12
+## WIP
 - Add support for automatic intro/full text image resizing (Thanks @nikosdion)
-- Use OS prefered colour scheme if theme hasn't been set
+- Add system preference to theme switcher
 - Fix "show theme switcher" parameter not working (Thanks @nikosdion)
+
+## 0.0.12
+- Use OS prefered colour scheme if theme hasn't been set
 - Fix CSS not being loaded if theme switcher was disabled
 - Fix styling for invalid field feedback